+++ CHANGELOG.md
@@ -7,6 +7,8 @@
   (for [#3267](https://github.com/pbek/QOwnNotes/issues/3267))
   - Please visit the [Debian installation page](https://www.qownnotes.org/installation/debian.html)
     for instructions and report issues
+- If you paste text as attachment, some file formats can now be detected if they
+  match a certain pattern (for [#3269](https://github.com/pbek/QOwnNotes/issues/3269))
 
 ## 25.5.1